Refactor imports from BannerGQLTypes to ./types.
