+++ src/api/users/index.ts
@@ -1,5 +1,5 @@
 import { GET_ME_QUERY } from "./constants";
-import { User } from "../../graphql/generator/BannerGQLTypes";
+import { User } from "../../types";
 import { client } from "../../lib/apollo";
 
 export const getMe = async (): Promise<User> => {
