+++ src/state/reduces/me.reducer.ts
@@ -1,8 +1,4 @@
-import {
-  UserRoles,
-  UserStatuses,
-  UserTypes,
-} from "../../graphql/generator/BannerGQLTypes";
+import { UserRoles, UserStatuses } from "../../types";
 import { Coordinates } from "../../types/_old/Coordinates";
 import { MeActions } from "../actions";
 
@@ -42,10 +38,10 @@ const initialMeState: IMeState = {
     lastName: "",
     profilePic: "",
     verified: false,
-    role: UserRoles.Basic,
+    role: UserRoles.User,
     screenName: "",
     status: UserStatuses.Active,
-    userType: UserTypes.Standard,
+    userType: "Standard",
     likes: [],
     media: [],
     posts: [],
