+++ src/types/_old/PostReply.ts
@@ -1,4 +1,4 @@
-import { Post } from "../../graphql/generator/BannerGQLTypes";
+import { Post } from "../graphql";
 import { User } from "./User";
 
 export type PostReply = {
